+++ Source/Bugsnag.h
@@ -34,7 +34,9 @@
 #import "BugsnagAppWithState.h"
 #import "BugsnagDevice.h"
 #import "BugsnagDeviceWithState.h"
+#import "BugsnagEndpointConfiguration.h"
 #import "BugsnagError.h"
+#import "BugsnagSession.h"
 #import "BugsnagStackframe.h"
 #import "BugsnagThread.h"