+++ Source/BugsnagEvent.m
@@ -12,6 +12,7 @@
 #include <sys/utsname.h>
 #endif
 
+#import <Foundation/Foundation.h>
 #import "BSGSerialization.h"
 #import "Bugsnag.h"
 #import "BugsnagCollections.h"
@@ -22,6 +23,11 @@
 #import "Private.h"
 #import "BSG_RFC3339DateTool.h"
 #import "BugsnagStacktrace.h"
+#import "BugsnagThread.h"
+#import "RegisterErrorData.h"
+#import "BugsnagSessionInternal.h"
+
+static NSString *const DEFAULT_EXCEPTION_TYPE = @"cocoa";
 
 // MARK: - Accessing hidden methods/properties
 
@@ -269,6 +275,8 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
  *  The release stage of the application
  */
 @property(readwrite, copy, nullable) NSString *releaseStage;
+
+@property NSArray *redactedKeys;
 @end
 
 @implementation BugsnagEvent
@@ -434,7 +442,7 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
     if ([BugsnagConfiguration isValidApiKey:apiKey]) {
         _apiKey = apiKey;
     }
-    
+
     // A malformed apiKey should not cause an error: the fallback global value
     // in BugsnagConfiguration will do to get the event reported.
     else {
@@ -545,7 +553,6 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
 
 - (NSDictionary *)toJson {
     NSMutableDictionary *event = [NSMutableDictionary dictionary];
-    NSMutableDictionary *metadata = [[[self metadata] toDictionary] mutableCopy];
 
     if (self.customException) {
         BSGDictSetSafeObject(event, @[ self.customException ], BSGKeyExceptions);
@@ -562,14 +569,17 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
     // Build Event
     BSGDictSetSafeObject(event, BSGFormatSeverity(self.severity), BSGKeySeverity);
     BSGDictSetSafeObject(event, [self serializeBreadcrumbs], BSGKeyBreadcrumbs);
-    BSGDictSetSafeObject(event, metadata, BSGKeyMetadata);
+
+    // add metadata
+    NSMutableDictionary *metadata = [[[self metadata] toDictionary] mutableCopy];
+    BSGDictSetSafeObject(event, [self sanitiseMetadata:metadata], BSGKeyMetadata);
 
     BSGDictSetSafeObject(event, [self.device toDictionary], BSGKeyDevice);
     BSGDictSetSafeObject(event, [self.app toDict], BSGKeyApp);
-    
+
     BSGDictSetSafeObject(event, [self context], BSGKeyContext);
     BSGDictInsertIfNotNil(event, self.groupingHash, BSGKeyGroupingHash);
-    
+
 
     BSGDictSetSafeObject(event, @(self.handledState.unhandled), BSGKeyUnhandled);
 
@@ -608,6 +618,52 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
     return event;
 }
 
+- (NSMutableDictionary *)sanitiseMetadata:(NSMutableDictionary *)metadata {
+    for (NSString *sectionKey in [metadata allKeys]) {
+        metadata[sectionKey] = [metadata[sectionKey] mutableCopy];
+        NSMutableDictionary *section = metadata[sectionKey];
+
+        if (section != nil) { // redact sensitive metadata values
+            for (NSString *objKey in [section allKeys]) {
+                section[objKey] = [self sanitiseMetadataValue:section[objKey] key:objKey];
+            }
+        }
+    }
+    return metadata;
+}
+
+- (id)sanitiseMetadataValue:(id)value key:(NSString *)key {
+    if ([self isRedactedKey:key]) {
+        return BSGKeyRedaction;
+    } else if ([value isKindOfClass:[NSDictionary class]]) {
+        NSMutableDictionary *nestedDict = [(NSDictionary *)value mutableCopy];
+
+        for (NSString *nestedKey in [nestedDict allKeys]) {
+            nestedDict[nestedKey] = [self sanitiseMetadataValue:nestedDict[nestedKey] key:nestedKey];
+        }
+        return nestedDict;
+    } else {
+        return value;
+    }
+}
+
+- (BOOL)isRedactedKey:(NSString *)key {
+    for (id obj in self.redactedKeys) {
+        if ([obj isKindOfClass:[NSString class]]) {
+            if ([key isEqualToString:obj]) {
+                return true;
+            }
+        } else if ([obj isKindOfClass:[NSRegularExpression class]]) {
+            NSRegularExpression *regex = obj;
+            NSRange range = NSMakeRange(0, [key length]);
+            if ([[regex matchesInString:key options:0 range:range] count] > 0) {
+                return true;
+            }
+        }
+    }
+    return false;
+}
+
 - (NSDictionary *)generateSessionDict {
     NSDictionary *events = @{
             @"handled": @(self.session.handledCount),
@@ -615,7 +671,7 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
     };
 
     NSDictionary *sessionJson = @{
-            BSGKeyId: self.session.sessionId,
+            BSGKeyId: self.session.id,
             @"startedAt": [BSG_RFC3339DateTool stringFromDate:self.session.startedAt],
             @"events": events
     };