+++ Tests/BugsnagSessionTrackerTest.m
@@ -24,6 +24,10 @@
 - (void)deletePersistedUserData;
 @end
 
+@interface BugsnagSessionTracker ()
+@property (strong, readwrite) BugsnagSession *currentSession;
+@end
+
 @interface BugsnagSessionTrackerTest : XCTestCase
 @property BugsnagConfiguration *configuration;
 @property BugsnagSessionTracker *sessionTracker;
@@ -45,7 +49,7 @@
     [self.sessionTracker startNewSession];
     BugsnagSession *session = self.sessionTracker.runningSession;
     XCTAssertNotNil(session);
-    XCTAssertNotNil(session.sessionId);
+    XCTAssertNotNil(session.id);
     XCTAssertTrue([[NSDate date] timeIntervalSinceDate:session.startedAt] < 1);
     XCTAssertFalse(session.autoCaptured);
 }
@@ -57,7 +61,7 @@
     BugsnagSession *session = self.sessionTracker.runningSession;
 
     XCTAssertNotNil(session);
-    XCTAssertNotNil(session.sessionId);
+    XCTAssertNotNil(session.id);
     XCTAssertTrue([[NSDate date] timeIntervalSinceDate:session.startedAt] < 1);
     XCTAssertEqual(session.user.name, @"Bill");
     XCTAssertEqual(session.user.userId, @"123");
@@ -71,7 +75,7 @@
     BugsnagSession *session = self.sessionTracker.runningSession;
 
     XCTAssertNotNil(session);
-    XCTAssertNotNil(session.sessionId);
+    XCTAssertNotNil(session.id);
     XCTAssertTrue([[NSDate date] timeIntervalSinceDate:session.startedAt] < 1);
     XCTAssertTrue(session.autoCaptured);
     XCTAssertNil(session.user.name);
@@ -86,7 +90,7 @@
     BugsnagSession *session = self.sessionTracker.runningSession;
 
     XCTAssertNotNil(session);
-    XCTAssertNotNil(session.sessionId);
+    XCTAssertNotNil(session.id);
     XCTAssertTrue([[NSDate date] timeIntervalSinceDate:session.startedAt] < 1);
     XCTAssertEqual(session.user.name, @"Bill");
     XCTAssertEqual(session.user.userId, @"123");
@@ -110,7 +114,7 @@
     [self.sessionTracker startNewSession];
 
     BugsnagSession *secondSession = self.sessionTracker.runningSession;
-    XCTAssertNotEqualObjects(firstSession.sessionId, secondSession.sessionId);
+    XCTAssertNotEqualObjects(firstSession.id, secondSession.id);
 }
 
 - (void)testIncrementCounts {
@@ -138,5 +142,32 @@
     XCTAssertEqual(1, session.handledCount);
     XCTAssertEqual(1, session.unhandledCount);
 }
+
+- (void)testOnSendBlockFalse {
+    self.configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
+    [self.configuration addOnSessionBlock:^BOOL(BugsnagSession *sessionPayload) {
+        return NO;
+    }];
+    self.sessionTracker = [[BugsnagSessionTracker alloc] initWithConfig:self.configuration
+                                                     postRecordCallback:nil];
+    [self.sessionTracker startNewSession];
+    XCTAssertNil(self.sessionTracker.currentSession);
+}
+
+- (void)testOnSendBlockTrue {
+    __block XCTestExpectation *expectation = [self expectationWithDescription:@"Session block is invoked"];
+
+    self.configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
+    [self.configuration addOnSessionBlock:^BOOL(BugsnagSession *sessionPayload) {
+        [expectation fulfill];
+        return YES;
+    }];
+    self.sessionTracker = [[BugsnagSessionTracker alloc] initWithConfig:self.configuration
+                                                     postRecordCallback:nil];
+    [self.sessionTracker startNewSession];
+    [self waitForExpectations:@[expectation] timeout:2];
+    XCTAssertNotNil(self.sessionTracker.currentSession);
+}
+
 
 @end