+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/AppDelegate.swift
@@ -52,7 +52,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // also that extra reports aren't erroneously sent in other conditions
         // when OOM reporting is enabled
         config.enabledErrorTypes.insert(.OOMs)
-        config.setEndpoints(notify: mockAPIPath, sessions: mockAPIPath)
+        config.endpoints = BugsnagEndpointConfiguration(notify: mockAPIPath, sessions: mockAPIPath)
         return config
     }